kamus-backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
psycopg2-binary
alembic
python-multipart
pydantic-settings